Вынесет список фильмов в отдельный компонент
